pyproject.toml
[project]
name = "mavis"
version = "0.1.0"
description = "Multi-view Action-to-Visual Image Synthesis"
readme = "README.md"
requires-python = ">=3.11,<3.12" 
dependencies = [
    "bpy>=5.0.1",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.6",
    "openai>=1.0.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "tenacity>=9.0.0",
    "mathutils>=3.3.0",
    "torch>=2.10.0",
    "diffusers>=0.36.0",
    "huggingface-hub>=1.4.1",
    "transformers>=5.1.0",
    "fal-client>=0.13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

